[GStreamer] Player audio sink needs to be initialized earlier https://bugs.webkit.org/show_bug.cgi?id=220462
Patch by Philippe Normand <pnormand@igalia.com> on 2021-01-08
Reviewed by Xabier Rodriguez-Calvar.

Currently the player initializes the audio sink in the load method but that's not enough in
situations where the player has an audiosourceprovider. So initialize it from the
constructor early on, to ensure that the audio provider sink bin is valid.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::MediaPlayerPrivateGStreamer):
(WebCore::MediaPlayerPrivateGStreamer::createGSTPlayBin):
* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h:

Canonical link: https://commits.webkit.org/232881@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271302 268f45cc-cd09-0410-ab3c-d52691b4dbfc