[GStreamer][EME][Thunder] Bail out on failed response message for challenges https://bugs.webkit.org/show_bug.cgi?id=220583
Reviewed by Philippe Normand.

When parsing the challenge message, sometimes the message length
is 0 and parsing fails. This should not happen in principle and
that's why we just asserted on that but it looks like the
framework is not as careful as expected here. Hence we need to
bail out to not crash in release mode.

* platform/graphics/gstreamer/eme/CDMThunder.cpp:
(WebCore::CDMInstanceSessionThunder::challengeGeneratedCallback):


Canonical link: https://commits.webkit.org/232997@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271437 268f45cc-cd09-0410-ab3c-d52691b4dbfc