Source/WebCore/PAL: [GPUProcess] Claim ownership of IOSurface backing for ImageBufferBackends in the WebProcess https://bugs.webkit.org/show_bug.cgi?id=220663
Reviewed by Geoffrey Garen.

Declare new IOSurface SPI.

* pal/spi/cocoa/IOSurfaceSPI.h:

Source/WebKit:
[GPUProcess] Claim ownership of Canvas IOSurface backing in the WebProcess
https://bugs.webkit.org/show_bug.cgi?id=220663

Reviewed by Geoffrey Garen.

Claim ownership of Canvas IOSurface backing in the WebProcess. If we don't do this, all
IOSurface memory gets attributed to the GPUProcess, making it a prime candidate for jetsam.
We need to attribute memory usage to the WebProcesses being served by the GPUProcess as much
as possible and this is a step in this direction.

Note that this patch currently only impacts Canvas IOSurface backing when "DOM Rendering in
GPUProcess" is disabled. When "DOM Rendering in GPUProcess" is enabled, we are not supposed
to use the IOSurface API in the WebProcess at all.

* WebProcess/GPU/graphics/cocoa/ImageBufferShareableMappedIOSurfaceBackend.cpp:
(WebKit::ImageBufferShareableMappedIOSurfaceBackend::create):


Canonical link: https://commits.webkit.org/233156@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271627 268f45cc-cd09-0410-ab3c-d52691b4dbfc