[iOS] Set background playback restriction for WebAudio https://bugs.webkit.org/show_bug.cgi?id=220879 <rdar://72949281>
Reviewed by Jer Noble.
Source/WebCore:

WebAudio does not play in the background on iOS both because audio doesn't play
when the category is set to `AVAudioSessionCategoryAmbient`, and because the
AudioContext is paused when the web process is forwarded the
`UIApplicationWillResignActiveNotification` notification.

This notification isn't forwarded when the WebView is hosted in SafariViewController
so the AudioContext is not paused, so it resumes playing when the application is
foregrounded even if the view controller is not in the foreground.

Fix this by setting the `BackgroundProcessPlaybackRestricted` restriction for
WebAudio sessions so the context is always paused in the background.

Test: media/webaudio-background-playback.html

* platform/audio/ios/MediaSessionManagerIOS.mm:
(WebCore::MediaSessionManageriOS::resetRestrictions):

LayoutTests:

* media/media-session-restrictions-expected.txt: Update expectations.
* media/webaudio-background-playback-expected.txt: Added.
* media/webaudio-background-playback.html: Added.


Canonical link: https://commits.webkit.org/233285@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271779 268f45cc-cd09-0410-ab3c-d52691b4dbfc