Nullptr dereference in SpeechRecognition::stop() https://bugs.webkit.org/show_bug.cgi?id=220869 <rdar://problem/73507085>
Patch by Sihui Liu <sihui_liu@appe.com> on 2021-01-25
Reviewed by Ryosuke Niwa.

Source/WebCore:

It is possible that SpeechRecognition is created with a Document having no frame and ActiveDOMObjects of the
Document are stopped. In this case, SpeechRecognition::stop() will be invoked after a SpeechRecognition, of
which m_connection is nullptr, is constructed.

Test: fast/speechrecognition/detached-iframe.html

* Modules/speech/SpeechRecognition.cpp:
(WebCore::SpeechRecognition::stop):

LayoutTests:

* fast/speechrecognition/detached-iframe-expected.txt: Added.
* fast/speechrecognition/detached-iframe.html: Added.

Canonical link: https://commits.webkit.org/233297@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271791 268f45cc-cd09-0410-ab3c-d52691b4dbfc