ASSERT NOT REACHED in WebCore::DisplayList::isDrawingItem https://bugs.webkit.org/show_bug.cgi?id=220948 <rdar://problem/73588734>
Reviewed by Tim Horton.

Source/WebCore:

Additionally, fix a bug that is also caught when running this API test under debug. In the case where an out of
line item failed to decode (and is thus invalid), we still attempt to invoke its destructor when we're done
iterating, under `DisplayList::iterator::clearCurrentItem()`; we should not be doing this in the case where
`m_isValid` is `false`, since we already know that `m_currentBufferForItem` either contains an item that has
already been destroyed, or contains garbage data.

* platform/graphics/displaylists/DisplayList.cpp:
(WebCore::DisplayList::DisplayList::iterator::clearCurrentItem):

Tools:

The API test `DisplayListTests.OutOfLineItemDecodingFailure` was intended to read from the display list copy
rather than the original list, in order to exercise the (intentional) decoding failure.

Instead, this test erroneously attempts to read from the original display list (which, importantly, doesn't have
a reading client) and ends up exercising the decoding failure anyways, but additionally hits a debug assertion
in the process. Simply fix this by reading out of the correct display list (and rename the original display list
in the process to make the test more clear).

* TestWebKitAPI/Tests/WebCore/cg/DisplayListTestsCG.cpp:
(TestWebKitAPI::TEST)


Canonical link: https://commits.webkit.org/233332@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271871 268f45cc-cd09-0410-ab3c-d52691b4dbfc