[Legacy Line Layout] Multiple inline boxes may stretch the line https://bugs.webkit.org/show_bug.cgi?id=220905
Reviewed by Antti Koivisto.

Source/WebCore:

This is a follow-up on r271110 where the top position of the inline box with the maximum ascent
gets adjusted to make sure it does not stick out of the line. This patch fixes the case when
multiple inline boxes have the same maximum ascent value.

* rendering/InlineFlowBox.cpp:
(WebCore::InlineFlowBox::computeLogicalBoxHeights):

LayoutTests:

* platform/ios/fast/block/basic/014-expected.txt:
* platform/mac/fast/block/basic/014-expected.txt:


Canonical link: https://commits.webkit.org/233336@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271875 268f45cc-cd09-0410-ab3c-d52691b4dbfc