[macOS] Re-enable ENABLE_SET_WEBCONTENT_PROCESS_INFORMATION_IN_NETWORK_PROCESS https://bugs.webkit.org/show_bug.cgi?id=221039 <rdar://problem/73665061> Source/WebCore:
Reviewed by Brent Fulgham.

The commit r271907 introduced a crash, since it denied access to the Launch Services daemon, which AppKit is using to get and set
application information, for example whether the app is foreground or not. To work around this, the WebCoreThemeWindow has been made
borderless, and the method +[NSWindow invalidateRestorableState] is overridden to do nothing. Running the default
+[NSWindow invalidateRestorableState] will exercise code which tries to reach out to the Launch Services daemon, and crashes if
unable to do so. This functionality should not be required in the WebContent process, and can be disabled.

* platform/mac/ThemeMac.mm:
(-[WebCoreThemeWindow invalidateRestorableState]):
(-[WebCoreThemeView window]):

Source/WebKit:

Reviewed by Brent Fulgham.

Work around crash introduced in r271907, by disabling persistent UI. Having persistent UI enabled will exercise code
which tries to connect with the Launch Services daemon, and crash if that fails.

* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess):

Source/WTF:

Reviewed by Brent Fulgham.

* wtf/PlatformEnableCocoa.h:


Canonical link: https://commits.webkit.org/233409@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@271965 268f45cc-cd09-0410-ab3c-d52691b4dbfc