REGRESSION(r262893): feComponentTransfer is not rendered correctly on Retina display https://bugs.webkit.org/show_bug.cgi?id=220576
Reviewed by Simon Fraser.

Source/WebCore:

convertImageDataToColorSpace() creates a temporary ImageBuffer with the
size of m_premultipliedImageResult. This ImageBuffer has to be scaled
with the filterScale(). But the size of m_premultipliedImageResult is
scaled with the filterScale(). This means, convertImageDataToColorSpace()
applies the filterScale() twice to the temporary ImageBuffer.

The fix is to create the temporary ImageBuffer with the unscaled size of
the input ImageData and scale it with the filterScale().

Test: fast/hidpi/filters-component-transfer.html

* platform/graphics/filters/FilterEffect.cpp:
(WebCore::FilterEffect::convertImageDataToColorSpace):
(WebCore::FilterEffect::copyPremultipliedResult):

LayoutTests:

* fast/hidpi/filters-component-transfer-expected.html: Added.
* fast/hidpi/filters-component-transfer.html: Added.


Canonical link: https://commits.webkit.org/233478@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@272059 268f45cc-cd09-0410-ab3c-d52691b4dbfc